Get my ideas for the Charm mode down somewhere concrete
I had these ideas on 04/25/202, in a DM with AWK, but I want to get them down here so we can start really planning it. I know it's not the most important part of the game, and people might barely use my weird note types, but hey, I'm a fan.